AG-15239 - Update trial license with zoom info (#4690)
