+++ external/ag-website-shared/src/components/google-tag-manager/GoogleTagManager.astro
@@ -1,0 +1,32 @@
+---
+interface Props {
+    googleTagManagerId: string;
+    googleTagManagerAuth: string;
+    googleTagManagerPreview: string;
+}
+
+const { googleTagManagerId, googleTagManagerAuth, googleTagManagerPreview } = Astro.props as Props;
+---
+
+<script
+    define:vars={{
+        googleTagManagerId,
+        googleTagManagerAuth,
+        googleTagManagerPreview,
+    }}
+>
+    (function (w, d, s, l, i) {
+        w[l] = w[l] || [];
+        w[l].push({ 'gtm.start': new Date().getTime(), event: 'gtm.js' });
+        var f = d.getElementsByTagName(s)[0],
+            j = d.createElement(s),
+            dl = l != 'dataLayer' ? '&l=' + l : '';
+        j.async = true;
+        j.src =
+            'https://www.googletagmanager.com/gtm.js?id=' +
+            i +
+            dl +
+            `&gtm_auth=${googleTagManagerAuth}&gtm_preview=${googleTagManagerPreview}&gtm_cookies_win=x`;
+        f.parentNode.insertBefore(j, f);
+    })(window, document, 'script', 'dataLayer', googleTagManagerId);
+</script>
