+++ packages/ag-charts-website/.env.build
@@ -10,3 +10,8 @@ PUBLIC_ALGOLIA_INDEX_PREFIX=ag-charts-dev
 PUBLIC_TRIAL_LICENCE_FORM_URL=https://us-central1-stripe-testing-19784.cloudfunctions.net/CreateLeadForTrialLK
 CHECK_LINKS=true
 PUBLIC_PACKAGE_VERSION=12.0.2
+
+# Google Tag Manager (staging configuration)
+PUBLIC_GTM_ID=GTM-T7JG534
+PUBLIC_GTM_AUTH=K7m5Yf0NyqFbLXvp-gPogw
+PUBLIC_GTM_PREVIEW=env-55
