+++ packages/ag-charts-website/.env.build.archive
@@ -8,3 +8,8 @@ PUBLIC_GRID_VERSION=34.0.2
 LIVE_SITEMAP_URL=https://www.ag-grid.com/charts/sitemap-0.xml
 PUBLIC_TRIAL_LICENCE_FORM_URL=https://us-central1-aggrid-ecommerce.cloudfunctions.net/CreateLeadForTrialLK
 PUBLIC_PACKAGE_VERSION=12.0.2
+
+# Google Tag Manager (same as production)
+PUBLIC_GTM_ID=GTM-T7JG534
+PUBLIC_GTM_AUTH=BZRbizt3P5eacrM4bCJO5g
+PUBLIC_GTM_PREVIEW=env-1
