+++ packages/ag-charts-website/src/content/metadata/metadata.json
@@ -1,6 +1,5 @@
 {
     "title": "AG Charts: JavaScript Charting Library for React, Angular and Vue",
     "description": "Create interactive, animated, high-performance charts & graphs with AG Charts JavaScript Charting library, from the makers of AG Grid. Built with JavaScript for React, Angular and Vue.",
-    "googleTagManagerId": "GTM-5GWLR86Q",
     "canonicalUrlBase": "https://www.ag-grid.com/charts"
 }
